requirements.txt
Flask
Pillow
pytesseract
sympy
gTTS
numpy
